# Improve performance of Parsers.parse(Number, ...)