This is an inherently type-unstable operation where we don't know exactly what type of thing we're going to parse, _even_ once we get into the `typeparser`, which isn't usually the case.

This is about 2x faster in the top-level case by switching to a functional style where you pass in an "applicator" function that gets applied to whatever objecdt we end up with. If a user calls the new `parsenumber` directly, they can get near-native performance with their own applicator.

Perf examples:

Before
```
julia> @btime Parsers.parse(Number, "9223372036854775808")
  294.762 ns (7 allocations: 240 bytes)
9223372036854775808

julia> @btime Parsers.parse(Number, "1.0")
  370.350 ns (9 allocations: 208 bytes)
1.0

julia> @btime Parsers.parse(Number, "1")
  249.674 ns (4 allocations: 112 bytes)
1
```

This PR:
```
julia> @btime Parsers.parse(Number, "1")
  123.481 ns (2 allocations: 48 bytes)
1

julia> @btime Parsers.parse(Number, "1.0")
  218.689 ns (4 allocations: 96 bytes)
1.0

julia> @btime Parsers.parse(Number, "9223372036854775808")
  153.579 ns (3 allocations: 80 bytes)
9223372036854775808
```